==> Heart_Disease_Project/notebooks/02_pca_analysis.py <==
# -*- coding: utf-8 -*-
"""02_pca_analysis.ipynb

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/18NFi56wGFhCWBXY_q101jiBW2gA3evI0
"""

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler

#Chargement des données netoyées
df = pd.read_csv("heart_clened.csv")
print("--"*40)
print( "\t\tApperçu Des Données " )
print("--"*40)
display(df.head())
print(df.shape)
print(df.info())

#Separation X / y
X = df.drop("target", axis=1)
y = df["target"]
scaler = StandardScaler()
X_scaled = scaler.fit_transform(X)
pca = PCA()
pca.fit(X_scaled)
explained_variance = pca.explained_variance_ratio_
cumulative_variance = explained_variance.cumsum()

print("--"*40)
print( "\t\tVisualisation de la Variance Expliquée " )
print("--"*40)
plt.figure(figsize=(12,6))
plt.bar(range(1, len(explained_variance)+1), explained_variance)
plt.plot(range(1, len(cumulative_variance)+1), cumulative_variance, marker='o')
plt.title("Variance expliquée & Variance cumulative")
plt.xlabel("Composante principale")
plt.ylabel("Ratio de variance expliquée")
plt.grid(True)
plt.show()

#determination du nombre optimal de composants
optimal_components = np.argmax(cumulative_variance >= 0.95) + 1
print("--"*40)
print( "\t\tVisualisation de la Variance Expliquée " )
print("--"*40)
print(f"Nombre optimal de composantes conservant 95% de la variance : {optimal_components}")

#PCA reduit avec le nombre optimal
pca_optimal = PCA(n_components=optimal_components)
X_pca = pca_optimal.fit_transform(X_scaled)
print("--"*40)
print( "\t\tPCA Reduit Avec Le Nombre Optimal " )
print("--"*40)
print("Nouvelle dimension Apres PCA : ", X_pca.shape)

# visualisation 2D PCA
print("--"*40)
print( "\t\tVisualisation 2D PCA " )
print("--"*40)
plt.figure(figsize=(10,7))
plt.scatter(X_pca[:, 0], X_pca[:, 1], c=y, cmap="coolwarm", alpha=0.8)
plt.xlabel("PC1")
plt.ylabel("PC2")
plt.title("Projection PCA (2 composantes principales)")
plt.colorbar(label="Target")
plt.show()

# Export des données transformées
pca_df = pd.DataFrame(X_pca, columns=[f"PC{i+1}" for i in range(optimal_components)])
pca_df["target"] = y.values
pca_df.to_csv("heart_pca.csv", index=False)
print("PCA Terminé Fichier Exporté: heart_pca.csv")
