==> Heart_Disease_Project/notebooks/07_export_model.py <==
# -*- coding: utf-8 -*-
"""07_export_model.ipynb

Automatically generated by Colab.

Original file is located at
    https://colab.research.google.com/drive/1I7JtArUBgo3Calw5a5ni3ArzQW61Xg0d

<h2> Exploration et Déploiement du Modèle </h2>
1. Charger le meilleur modèle<br>
2. Construire un pipeline complet (scaler + modèle)<br>
3. L’exporter proprement en .pkl<br>
4. Garantir la reproductibilité<br>
"""

import pandas as pd
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
import joblib

"""<h2> Chargement des données Du Sprint 3</h2>
heart_selected_features.csv
"""

df = pd.read_csv("heart_selected_features.csv")
X = df.drop("target", axis=1)
y = df["target"]

"""<h2>Chargement du meilleur Modele Sprint 6</h2>"""

best_model = joblib.load("final_model.pkl")
best_model

"""<h2> Construction de Pipeline Complet (Scaler --> Modéle) </h2>"""

pipeline = Pipeline(steps=[
    ("scaler", StandardScaler()),
    ("model", best_model)
])

"""<h2> Entrainement le pipeline complet </h2>"""

pipeline.fit(X, y)

"""<h2> Exportation Pipeline en .pkl"""

output_path = "final_heart_disease_model.pkl"
joblib.dump(pipeline, output_path)
print("Modèle exporté avec succès : ", output_path)
